pyproject.toml
[tool.poetry]
name = "mobile-api"
version = "0.1.0"
description = ""
authors = ["Maxud-M <evistix28@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "4.2.7"
djangorestframework = "^3.14.0"
pillow = "^10.2.0"
django-phonenumber-field = "^7.3.0"
pytils = "^0.4.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
